Added main menu and music control
Added the main menu, still needs some work like changing name of title. Also when using the audio control, just copy and paste it into the scene you make and it should work just fine. Also I added some audio files for shooting for the ranged enemy ai and I added hit sounds for when the player or ai gets hit. They aren't put on the ai yet but it's easy to add.